refactor out graph generation
